requirements.txt
annotated-types==0.7.0
anyio==4.11.0
arabic-reshaper==3.0.0
asgiref==3.10.0
asn1crypto==1.5.1
cachetools==6.2.2
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.3
cssselect2==0.8.0
Django==4.2.26
django-colorfield==0.14.0
django-jazzmin==3.0.1
et_xmlfile==2.0.0
exceptiongroup==1.3.0
freetype-py==2.5.1
google-auth==2.43.0
google-genai==1.47.0
h11==0.16.0
html5lib==1.1
httpcore==1.0.9
httpx==0.28.1
idna==3.11
lxml==6.0.2
numpy==2.0.2
openpyxl==3.1.5
oscrypto==1.3.0
packaging==25.0
pandas==2.3.3
pillow==11.3.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycairo==1.28.0
pycparser==2.23
pydantic==2.12.4
pydantic_core==2.41.5
pyHanko==0.31.0
pyhanko-certvalidator==0.29.0
pypdf==6.2.0
python-bidi==0.6.7
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
reportlab==4.4.4
requests==2.32.5
rlPyCairo==0.4.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
svglib==1.6.0
tenacity==9.1.2
tinycss2==1.4.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
uritools==5.0.0
urllib3==2.5.0
webencodings==0.5.1
websockets==15.0.1
xhtml2pdf==0.2.17
